add abs value define width, delete double from fir_filter

+++ src/adder.m
@@ -4,10 +4,10 @@ function [y, overflow] = adder(a,b,N)
     y = a + b;
 
     if (y > (2^(N-1)-1) | y < -2^(N-1))
-        overflow = 1;
+        overflow = int8(1);
         disp('overflow add detected');
     else
-        overflow = 0;
+        overflow = int8(0);
     end
